# openapi/speaking-history.yaml
openapi: 3.0.3
info:
  title: FlashMind Speaking 歷史與 CLI 契約
  version: 1.0.0
paths:
...
  /auth/cli/authorizations:
    post:
      operationId: createCliLoginAuthorization
      tags:
        - CliAuth
      summary: 建立短期 CLI 登入授權
      security: []
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/CliAuthorizationStarted"
                required:
                  - data
        "400":
          description: 格式錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "401":
          description: 未登入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "403":
          description: 禁止存取
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "404":
          description: 找不到資源
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "409":
          description: 內容或版本衝突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "413":
          description: 資料超過限制
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "422":
          description: 驗證不符
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "429":
          description: 請求過於頻繁
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliAuthorizationCreate"
  /auth/cli/authorizations/{id}/approve:
    post:
      operationId: approveCliLoginAuthorization
      tags:
        - CliAuth
      summary: 瀏覽器確認 CLI 帳號及配對碼
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/CliAuthorizationStatus"
                required:
                  - data
        "400":
          description: 格式錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "401":
          description: 未登入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "403":
          description: 禁止存取
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "404":
          description: 找不到資源
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "409":
          description: 內容或版本衝突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "413":
          description: 資料超過限制
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "422":
          description: 驗證不符
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "429":
          description: 請求過於頻繁
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "410":
          description: 授權已過期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliAuthorizationApprove"
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
  /auth/cli/authorizations/{id}/exchange:
    post:
      operationId: exchangeCliLoginAuthorization
      tags:
        - CliAuth
      summary: 由發起端單次兌換 session cookie
      security: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    $ref: "#/components/schemas/CliAuthorizationStatus"
                required:
                  - data
          headers:
            Set-Cookie:
              description: 授權完成時回傳新的 HttpOnly session；不得記錄或放入 JSON
              schema:
                type: string
        "400":
          description: 格式錯誤
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "401":
          description: 未登入
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "403":
          description: 禁止存取
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "404":
          description: 找不到資源
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "409":
          description: 內容或版本衝突
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "413":
          description: 資料超過限制
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "422":
          description: 驗證不符
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "429":
          description: 請求過於頻繁
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
        "410":
          description: 授權已過期
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeakingHistoryError"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CliAuthorizationExchange"
      parameters:
        - name: id
          in: path
          required: true
          schema: *id001
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    SpeakingValidationIssue:
      type: object
      additionalProperties: false
      properties:
        path:
          type: string
        code:
          type: string
        message:
          type: string
      required:
        - path
        - code
        - message
...
    CliAuthorizationCreate:
      type: object
      additionalProperties: false
      properties:
        verifierHash:
          type: string
          pattern: ^[a-f0-9]{64}$
      required:
        - verifierHash
    CliAuthorizationStarted:
      type: object
      additionalProperties: false
      properties:
        authorizationId: *id001
        verificationUrl:
          type: string
          format: uri
        pairingCode: *id001
        expiresAt: *id004
        pollIntervalMs:
          type: integer
      required:
        - authorizationId
        - verificationUrl
        - pairingCode
        - expiresAt
        - pollIntervalMs
    CliAuthorizationApprove:
      type: object
      additionalProperties: false
      properties:
        pairingCode: *id001
        decision:
          type: string
          enum:
            - approve
            - deny
        expectedUserId:
          type: string
          minLength: 1
          maxLength: 200
          description: 頁面確認的帳號；須與當前 session 相同，避免另一分頁切換帳號後誤寫
      required:
        - pairingCode
        - decision
        - expectedUserId
    CliAuthorizationExchange:
      type: object
      additionalProperties: false
      properties:
        verifier:
          type: string
          minLength: 43
          maxLength: 128
      required:
        - verifier
    CliAuthorizationStatus:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum:
            - pending
            - approved
            - denied
        expiresAt: *id004
        userId:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required:
        - status
        - expiresAt
        - userId
        - email
    SpeakingHistoryError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                $ref: "#/components/schemas/SpeakingValidationIssue"
          required:
            - code
            - message
      required:
        - error