# packages/api-client/src/generated/model/cliAuthorizationApprove.ts
/**
 * FlashMind API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CliAuthorizationApprove {
  pairingCode: string;
  decision: CliAuthorizationApprove.DecisionEnum;
  /**
   * 頁面確認的帳號；須與當前 session 相同，避免另一分頁切換帳號後誤寫
   */
  expectedUserId: string;
}
export namespace CliAuthorizationApprove {
  export const DecisionEnum = {
    Approve: "approve",
    Deny: "deny",
  } as const;
  export type DecisionEnum = (typeof DecisionEnum)[keyof typeof DecisionEnum];
}