# packages/api-client/src/generated/model/cliAuthorizationStatus.ts
/**
 * FlashMind API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CliAuthorizationStatus {
  status: CliAuthorizationStatus.StatusEnum;
  expiresAt: string;
  userId: string | null;
  email: string | null;
}
export namespace CliAuthorizationStatus {
  export const StatusEnum = {
    Pending: "pending",
    Approved: "approved",
    Denied: "denied",
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
}